MOBI-489 Showcase Updates - fix NPE on shutdown in Tomcat6
git-svn-id: http://anonsvn.icesoft.org/repo/icemobile/trunk/icemobile@33025 1de00972-ac04-4911-a46b-06793c8d78d9

+++ samples/mobileshowcase/src/main/java/org/icemobile/samples/mobileshowcase/view/examples/device/camera/CameraBean.java
@@ -122,9 +122,11 @@ public class CameraBean extends ExampleImpl<CameraBean> implements
 
     @PreDestroy
     public void disposeResources(){
-        boolean success = cameraFile.delete();
-        if (!success && logger.isLoggable(Level.FINE)){
-            logger.fine("Could not dispose of media file" + cameraFile.getAbsolutePath());
+        if( cameraFile != null ){
+            boolean success = cameraFile.delete();
+            if (!success && logger.isLoggable(Level.FINE)){
+                logger.fine("Could not dispose of media file" + cameraFile.getAbsolutePath());
+            }
         }
     }
